Break server code up into logical parts
This will make future improvements easier to manage, and improves code readability.
